+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
   * Add `FailoverAdapterFactory`
 
+### 🐛 Bug fixes
+
+  * Do not throw `FunctionsNotAllowedException` in `supports` method of `AwsS3AdapterFactory` and `OpenStackSwiftAdapterFactory` 
+
 ## v0.1.0 (August 30, 2021)
 
 First version.